integration-tests/requirements.txt
assertpy==1.*
behave
black
environs
faker
isort
mypy
reportportal-behave-client==1.*
requests==2.*
